Attempt to fix intermittent issues with curve25519-sha2.
The Q_C value was not properly zero-extended to the full 32-byte length
before being inserted into the SSH_MSG_KEX_ECDH_INIT packet, as well as
for use in the hash H computation.

Additionally, stop trying to clear sign bit of Q values: this appears to
be unnecessary (or is handled by the JCE X25519 implementation).

Finally, validate the length of Q_S sent by the server in order to
comply with section 3 of RFC 8731.